aosp: Get deps for Mesa3D build
Mesa3D is looks to be more interesting
for android usage out of x86 aosp based images.

Mesa3D vulkan runtime now have support for
android native buffer, that is necessary for skiavk
it's allow for some drivers like turnip works in android good
as well and being as alternative for qcom proprietary driver.

Google is now working with gfxstream on mesa3d that is very
initial state.
In future mesa3d can be more used as native drivers out of
x86 images.